chore(change-project-structure): Change project structure
This commit rearranges files into different folders for better folder structure
