+++ .travis.yml
@@ -1,9 +1,14 @@
 language: node_js
 node_js:
   - "stable"
+before_script:
+  - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
+  - chmod +x ./cc-test-reporter
+  - ./cc-test-reporter before-build
 script:
   - npm test -- --coverage # Include coverage when running tests
   - npm run build
 after_script: # Upload coverage reports 
   - COVERALLS_REPO_TOKEN=$coveralls_repo_token npm run coveralls
+  - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT -r a1f2bb9c030110934ee69446a387d805d4a8d4fdfe0c1dbf3898e2ce0f36a813
