+++ package.json
@@ -41,13 +41,16 @@
     "eslint-plugin-prettier": "^2.6.1",
     "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^3.1.0",
+    "fetch-mock": "^6.5.0",
     "jest": "^23.2.0",
     "prettier": "^1.13.7",
     "react-test-renderer": "^16.4.1",
     "redux-mock-store": "^1.5.3"
   },
   "jest": {
-    "snapshotSerializers": ["enzyme-to-json/serializer"]
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
   },
   "proxy": "http://127.0.0.1:5000"
 }
