+++ src/components/business/edit.js
@@ -5,42 +5,41 @@ import { withRouter } from 'react-router-dom';
 import decode from 'jwt-decode';
 import {NotificationManager} from 'react-notifications';
 
+import Weconnect from '../../store';
+
 import 'react-notifications/lib/notifications.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'jquery/dist/jquery.min.js';
 import 'bootstrap/dist/js/bootstrap.min.js';
-import {registerBusiness} from '../actions/registerBusinessActions';
-import AuthNavigationBar from './authNavigationBar';
+import { editBusiness } from '../../actions/getOneBusinessActions';
+import AuthNavigationBar from '../navBar/authNavigationBar';
 
-class BusinessRegister extends Component {
+class EditBusiness extends Component {
 
   componentDidMount=()=>{
-
     var userToken = sessionStorage.getItem("access_token");
     const userDecoded = decode(userToken);
-    if ((userToken !== null) && (userDecoded.exp > Date.now() / 1000)) {
-      this.props.history.push("/register-business")
-    }
-    else{
+    if ((userToken === null) || (userDecoded.exp < Date.now() / 1000)) {
+      console.log(this.props)
       this.props.history.push("/")
     }
   }
-  
+
   componentWillReceiveProps(receivedProp){
     console.log(receivedProp)
-    if(receivedProp.registerBizMessage.message){
-      if(receivedProp.registerBizMessage.message === "Business registered!"){
-        NotificationManager.success("Business registered!","", 5000);
+    if(receivedProp.editBusinessMessage.message){
+      if(receivedProp.editBusinessMessage.message === "Business successfully Updated!"){
+        NotificationManager.success("Business successfully Updated!","", 5000);
         this.props.history.push("/businesses")
       }else{
         NotificationManager.error(receivedProp.registerBizMessage.message,"", 5000);
-        console.log(receivedProp.registerBizMessage.message)
+        console.log(receivedProp.editBusinessMessage.message)
       }
     }
     
   }
 
-  registerBizDataStringify = (object) =>{
+  updateBizDataStringify = (object) =>{
     let simpleObj={};
         for (let prop in object){
             if (!object.hasOwnProperty(prop)){
@@ -55,7 +54,7 @@ class BusinessRegister extends Component {
 
   }
 
-  registerBiz=(event)=>{
+  editOneBusiness=(event)=>{
     event.preventDefault();
     let businessData={
       business_name:event.target.elements.businessName.value,
@@ -63,10 +62,21 @@ class BusinessRegister extends Component {
       location:event.target.elements.location.value,
       category:event.target.elements.category.value
     };
-    this.props.registerBusiness(this.registerBizDataStringify(businessData))
+    const bizId = Weconnect.getState().getBusiness.getBusinessMessage.business.id
+    this.props.editBusiness(bizId, this.updateBizDataStringify(businessData))
   }
 
   render() {
+    console.log(Weconnect.getState().getBusiness.getBusinessMessage.business)
+    if(Weconnect.getState().getBusiness.getBusinessMessage.business){
+      var businessName = Weconnect.getState().getBusiness.getBusinessMessage.business.BusinessName;
+      var businessProfile = Weconnect.getState().getBusiness.getBusinessMessage.business.BusinessProfile;
+      var businessLocation = Weconnect.getState().getBusiness.getBusinessMessage.business.Location;
+      var businessCategory = Weconnect.getState().getBusiness.getBusinessMessage.business.Category;
+      }
+    else{
+      this.props.history.push(`/businesses/${Weconnect.getState().getBusiness.getBusinessMessage.business.id}`)
+    }
     return (
       <div className="businessRegister">
         <AuthNavigationBar/>
@@ -74,14 +84,14 @@ class BusinessRegister extends Component {
        <div className="container">
         <div className="row">
             <div className="col bg-info">
-              <h4 style={{textAlign: 'center', color: '#fff', fontWeight: 'bolder'}}><br />Register a business</h4><br/>
-              <form onSubmit={this.registerBiz}>
+              <h4 style={{textAlign: 'center', color: '#fff', fontWeight: 'bolder'}}><br />Edit business</h4><br/>
+              <form onSubmit={this.editOneBusiness}>
                 <div className="form-group">
-                  <input type="text" className="form-control" id="businessName" placeholder="Enter Business Name" name="businessName" required="required" />
+                  <input type="text" className="form-control" defaultValue={businessName} id="businessName" placeholder="Enter Business Name" name="businessName" required="required" />
                 </div>
                 <div className="form-group">
                   <select className="form-control" id="location" name="location" required="required">
-                    <option value="">Business Location</option> 
+                    <option defaultValue="{businessLocation}">{businessLocation}</option> 
                     <option value="Mbarara">Mbarara</option>
                     <option value="Kampala">Kampala</option>
                     <option value="Gulu">Gulu</option>
@@ -91,7 +101,7 @@ class BusinessRegister extends Component {
                   </select>
                   <br />
                   <select className="form-control" id="category" name="category" required="required">
-                    <option value="">Business Category</option>  
+                    <option defaultValue="{businessCategory}">{businessCategory}</option>  
                     <option value="Consulting">Consulting</option>
                     <option value="Telecommunications">Telecommunications</option>
                     <option value="Food and Beverages">Food and Beverages</option>
@@ -101,10 +111,10 @@ class BusinessRegister extends Component {
                   </select>
                 </div>
                 <div className="form-group">
-                  <textarea className="form-control" placeholder="Enter Description of the business" id="descr" name="descr" required="required" />
+                  <textarea className="form-control" defaultValue={businessProfile} placeholder="Enter Description of the business" id="descr" name="descr" required="required" />
                 </div>
                 <br />
-                <button type="submit" className="btn btn-primary" style={{width: '25%', marginLeft: '25%', backgroundColor: '#fff', borderColor: '#14a2b8', color: '#14a2b8'}}>Register</button>
+                <button type="submit" className="btn btn-primary" style={{width: '25%', marginLeft: '25%', backgroundColor: '#fff', borderColor: '#14a2b8', color: '#14a2b8'}}>Update</button>
                 <a className="btn btn-warning" style={{width: '25%', borderColor: '#14a2b8', color: '#fff'}} href="/businesses" role="button">Cancel</a>
               </form><br /><br />
             </div>
@@ -115,13 +125,13 @@ class BusinessRegister extends Component {
   }
 }
 
-BusinessRegister.propTypes = {
-  registerBizMessage: PropTypes.object,
-  registerBusiness: PropTypes.func.isRequired
+EditBusiness.propTypes = {
+  editBusiness:PropTypes.func.isRequired,
+  editBusinessMessage: PropTypes.object.isRequired,
 }
 
-const mapStateToProps = state => ({
-  registerBizMessage: state.registerBusiness.registerBusinessMessage
+const mapStateToProps = (state, ownProps) => ({
+  editBusinessMessage:state.getBusiness.editBusinessMessage,
 });
 
-export default withRouter(connect(mapStateToProps, {registerBusiness})(BusinessRegister));
+export default withRouter(connect(mapStateToProps, {editBusiness})(EditBusiness));
